requirements_cursor_smc.txt
# SMC Trading Bot Requirements

# Core
MetaTrader5>=5.0.4
pandas>=1.5.0
numpy>=1.23.0
PyYAML>=6.0

# Optional but recommended
python-dateutil>=2.8.2
pytz>=2022.1
